+++ packages/core/src/routes/session/forgot-password.test.ts
@@ -97,20 +97,18 @@ describe('session -> forgotPasswordRoutes', () => {
       const response = await sessionRequest
         .post(`${forgotPasswordRoute}/sms/verify-passcode`)
         .send({ phone: '13000000000', code: '1234' });
-      expect(response.statusCode).toEqual(200);
-      expect(response.body).toHaveProperty('redirectTo');
+      expect(response.statusCode).toEqual(204);
       expect(interactionResult).toHaveBeenCalledWith(
         expect.anything(),
         expect.anything(),
         expect.objectContaining({
-          login: { accountId: 'id' },
           forgotPassword: {
+            userId: 'id',
             expiresAt: dayjs(fakeTime)
               .add(forgotPasswordVerificationTimeout, 'second')
               .toISOString(),
           },
-        }),
-        expect.anything()
+        })
       );
       jest.useRealTimers();
     });
@@ -150,20 +148,18 @@ describe('session -> forgotPasswordRoutes', () => {
       const response = await sessionRequest
         .post(`${forgotPasswordRoute}/email/verify-passcode`)
         .send({ email: 'a@a.com', code: '1234' });
-      expect(response.statusCode).toEqual(200);
-      expect(response.body).toHaveProperty('redirectTo');
+      expect(response.statusCode).toEqual(204);
       expect(interactionResult).toHaveBeenCalledWith(
         expect.anything(),
         expect.anything(),
         expect.objectContaining({
-          login: { accountId: 'id' },
           forgotPassword: {
+            userId: 'id',
             expiresAt: dayjs(fakeTime)
               .add(forgotPasswordVerificationTimeout, 'second')
               .toISOString(),
           },
-        }),
-        expect.anything()
+        })
       );
       jest.useRealTimers();
     });
@@ -188,8 +184,7 @@ describe('session -> forgotPasswordRoutes', () => {
     it('assign result and redirect', async () => {
       interactionDetails.mockResolvedValueOnce({
         result: {
-          login: { accountId: 'id' },
-          forgotPassword: { expiresAt: dayjs().add(1, 'day').toISOString() },
+          forgotPassword: { userId: 'id', expiresAt: dayjs().add(1, 'day').toISOString() },
         },
       });
       const response = await sessionRequest
@@ -219,8 +214,7 @@ describe('session -> forgotPasswordRoutes', () => {
     it('should throw when `forgotPassword.expiresAt` is not string', async () => {
       interactionDetails.mockResolvedValueOnce({
         result: {
-          login: { accountId: 'id' },
-          forgotPassword: { expiresAt: 0 },
+          forgotPassword: { userId: 'id', expiresAt: 0 },
         },
       });
       const response = await sessionRequest
@@ -232,8 +226,7 @@ describe('session -> forgotPasswordRoutes', () => {
     it('should throw when `expiresAt` is not a valid date string', async () => {
       interactionDetails.mockResolvedValueOnce({
         result: {
-          login: { accountId: 'id' },
-          forgotPassword: { expiresAt: 'invalid date string' },
+          forgotPassword: { userId: 'id', expiresAt: 'invalid date string' },
         },
       });
       const response = await sessionRequest
@@ -245,8 +238,7 @@ describe('session -> forgotPasswordRoutes', () => {
     it('should throw when verification expires', async () => {
       interactionDetails.mockResolvedValueOnce({
         result: {
-          login: { accountId: 'id' },
-          forgotPassword: { expiresAt: dayjs().subtract(1, 'day').toISOString() },
+          forgotPassword: { userId: 'id', expiresAt: dayjs().subtract(1, 'day').toISOString() },
         },
       });
       const response = await sessionRequest
@@ -258,8 +250,7 @@ describe('session -> forgotPasswordRoutes', () => {
     it('should throw when new password is the same as old one', async () => {
       interactionDetails.mockResolvedValueOnce({
         result: {
-          login: { accountId: 'id' },
-          forgotPassword: { expiresAt: dayjs().add(1, 'day').toISOString() },
+          forgotPassword: { userId: 'id', expiresAt: dayjs().add(1, 'day').toISOString() },
         },
       });
       mockArgon2Verify.mockResolvedValueOnce(true);
@@ -272,8 +263,7 @@ describe('session -> forgotPasswordRoutes', () => {
     it('should redirect when there was no old password', async () => {
       interactionDetails.mockResolvedValueOnce({
         result: {
-          login: { accountId: 'id' },
-          forgotPassword: { expiresAt: dayjs().add(1, 'day').toISOString() },
+          forgotPassword: { userId: 'id', expiresAt: dayjs().add(1, 'day').toISOString() },
         },
       });
       findUserById.mockResolvedValueOnce({