+++ packages/ui/src/apis/forgot-password.ts
@@ -18,8 +18,8 @@ export const sendForgotPasswordSmsPasscode = async (phone: string) => {
   return { success: true };
 };
 
-export const verifyForgotPasswordSmsPasscode = async (phone: string, code: string) =>
-  api
+export const verifyForgotPasswordSmsPasscode = async (phone: string, code: string) => {
+  await api
     .post(`${forgotPasswordApiPrefix}/sms/verify-passcode`, {
       json: {
         phone,
@@ -27,6 +27,9 @@ export const verifyForgotPasswordSmsPasscode = async (phone: string, code: strin
       },
     })
     .json<Response>();
+
+  return { success: true };
+};
 
 export const sendForgotPasswordEmailPasscode = async (email: string) => {
   await api
@@ -40,8 +43,8 @@ export const sendForgotPasswordEmailPasscode = async (email: string) => {
   return { success: true };
 };
 
-export const verifyForgotPasswordEmailPasscode = async (email: string, code: string) =>
-  api
+export const verifyForgotPasswordEmailPasscode = async (email: string, code: string) => {
+  await api
     .post(`${forgotPasswordApiPrefix}/email/verify-passcode`, {
       json: {
         email,
@@ -50,9 +53,15 @@ export const verifyForgotPasswordEmailPasscode = async (email: string, code: str
     })
     .json<Response>();
 
-export const resetPassword = async (password: string) =>
-  api
+  return { success: true };
+};
+
+export const resetPassword = async (password: string) => {
+  await api
     .post(`${forgotPasswordApiPrefix}/reset`, {
       json: { password },
     })
     .json<Response>();
+
+  return { success: true };
+};